chat/pyproject.toml
[tool.poetry]
name = "xiaoxiao"
version = "0.1.0"
description = ""
authors = ["XZhang <xiaozhang.uh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
marvin = "^2.1.5"
langchain = "^0.1.7"
langchain-openai = "^0.0.6"
langchain-pinecone = "^0.0.2"

chainlit = "^1.0.200"
tenacity = "^8.2.3"
langchainhub = "^0.1.14"
pinecone-client = "^3.0.3"

jupyter = "^1.0.0"
ipykernel = "^6.29.2"
tabulate = "^0.9.0"
pyspellchecker = "^0.8.1"
questionary = "^2.0.1"
halo = "^0.0.31"
ruff = "^0.2.2"
pandas = "^2.2.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
